fix(vercel): move api helpers out of serverless function directory
