+++ api/telegram.ts
@@ -6,8 +6,8 @@ import type {
     InlineKeyboardMarkup,
     AssignmentDoc,
     AssignmentWithDate
-} from './types.js';
-import { ASSIGNMENT_STATUS, ASSIGNMENT_PRIORITY } from './types.js';
+} from '../server/telegram/types.js';
+import { ASSIGNMENT_STATUS, ASSIGNMENT_PRIORITY } from '../server/telegram/types.js';
 
 if (!admin.apps || admin.apps.length === 0) {
     let privateKey = process.env.FIREBASE_PRIVATE_KEY || '';
