+++ server/chat/error-utils.ts
